Implementing permissions into model display managers

+++ models/entity/model.php
@@ -36,49 +36,68 @@ class model {
 		$this->loadRecord($parent);
 		$ret = false;
 		switch($action) {
+			case "read":
+				if($user->checkPermission($this->table, 'read', false)) {
+					$this->displayModelList();		
+				} else {
+					echo "You do not have permissions to '<strong>read</strong>' records for " . $this->table . ".<br />";
+				}		
+				break;
 			case "insert":
-				//Determine if the form has been submitted
-				if(isset($_POST['saveChanges'])) {
-					// User has posted the article edit form: save the new article
+				if($user->checkPermission($this->table, 'view', false)) {
+					//Determine if the form has been submitted
+					if(isset($_POST['saveChanges'])) {
+						// User has posted the article edit form: save the new article
+						
+						$this->storeFormValues($_POST);
+						
+						$result = $this->insert();
 					
-					$this->storeFormValues($_POST);
-					
-					$result = $this->insert();
-				
-					if(!$result) {
-						$this->buildEditForm($parent, $child, $user);
+						if(!$result) {
+							$this->buildEditForm($parent, $child, $user);
+						} else {
+							$this->buildEditForm(getLastField($this->conn,$this->table, "id"), $child, $user);
+							$this->log->trackChange($this->table, 'add',$user->getId(),$user->getLoginname(), $this->logField . " added");
+						}
 					} else {
-						$this->buildEditForm(getLastField($this->conn,$this->table, "id"), $child, $user);
-						$this->log->trackChange($this->table, 'add',$user->getId(),$user->getLoginname(), $this->logField . " added");
+						// User has not posted the template edit form yet: display the form
+						$this->buildEditForm($parent, $child, $user);
 					}
 				} else {
-					// User has not posted the template edit form yet: display the form
-					$this->buildEditForm($parent, $child, $user);
+					echo "You do not have permissions to '<strong>insert</strong>' records for " . $this->table . ".<br />";
 				}
 				break;
 			case "update":
-				//Determine if the form has been submitted
-				if(isset($_POST['saveChanges'])) {
-					// User has posted the article edit form: save the new article
+				if($user->checkPermission($this->table, 'update', false)) {
+					//Determine if the form has been submitted
+					if(isset($_POST['saveChanges'])) {
+						// User has posted the article edit form: save the new article
+							
+						$this->storeFormValues($_POST);
 						
-					$this->storeFormValues($_POST);
-					
-					$result = $this->update($parent);
-					//Re-build the page creation form once we are done
-					$this->buildEditForm($parent, $child, $user);
-
-					if($result) {
-						$this->log->trackChange($this->table, 'update',$user->getId(),$user->getLoginname(), $this->logField . " updated");
+						$result = $this->update($parent);
+						//Re-build the page creation form once we are done
+						$this->buildEditForm($parent, $child, $user);
+	
+						if($result) {
+							$this->log->trackChange($this->table, 'update',$user->getId(),$user->getLoginname(), $this->logField . " updated");
+						}
+					} else {
+						// User has not posted the template edit form yet: display the form
+						$this->buildEditForm($parent, $child, $user);
 					}
 				} else {
-					// User has not posted the template edit form yet: display the form
-					$this->buildEditForm($parent, $child, $user);
+					echo "You do not have permissions to '<strong>update</strong>' records for " . $this->table . ".<br />";
 				}
 				break;
 			case "delete":
-				$this->delete($parent);
-				$ret = true;
-				$this->log->trackChange($this->table, 'delete',$user->getId(),$user->getLoginname(), $this->logField . " deleted");
+				if($user->checkPermission($this->table, 'delete', false)) {
+					$this->delete($parent);
+					$ret = true;
+					$this->log->trackChange($this->table, 'delete',$user->getId(),$user->getLoginname(), $this->logField . " deleted");
+				} else {
+					echo "You do not have permissions to '<strong>delete</strong>' records for " . $this->table . ".<br />";
+				}
 				break;
 			default:
 				echo "Error with " . $this->table . " manager<br /><br />";
